+++ x/erc20/types/msg.go
@@ -1,7 +1,7 @@
 package types
 
 import (
-	sdkerrors "cosmossdk.io/errors"
+	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
@@ -44,14 +44,14 @@ func (msg MsgConvertCoin) ValidateBasic() error {
 	}
 
 	if !msg.Coin.Amount.IsPositive() {
-		return sdkerrors.Wrapf(errortypes.ErrInvalidCoins, "cannot mint a non-positive amount")
+		return errorsmod.Wrapf(errortypes.ErrInvalidCoins, "cannot mint a non-positive amount")
 	}
 	_, err := sdk.AccAddressFromBech32(msg.Sender)
 	if err != nil {
-		return sdkerrors.Wrap(err, "invalid sender address")
+		return errorsmod.Wrap(err, "invalid sender address")
 	}
 	if !common.IsHexAddress(msg.Receiver) {
-		return sdkerrors.Wrapf(errortypes.ErrInvalidAddress, "invalid receiver hex address %s", msg.Receiver)
+		return errorsmod.Wrapf(errortypes.ErrInvalidAddress, "invalid receiver hex address %s", msg.Receiver)
 	}
 	return nil
 }
@@ -86,17 +86,17 @@ func (msg MsgConvertERC20) Type() string { return TypeMsgConvertERC20 }
 // ValidateBasic runs stateless checks on the message
 func (msg MsgConvertERC20) ValidateBasic() error {
 	if !common.IsHexAddress(msg.ContractAddress) {
-		return sdkerrors.Wrapf(errortypes.ErrInvalidAddress, "invalid contract hex address '%s'", msg.ContractAddress)
+		return errorsmod.Wrapf(errortypes.ErrInvalidAddress, "invalid contract hex address '%s'", msg.ContractAddress)
 	}
 	if !msg.Amount.IsPositive() {
-		return sdkerrors.Wrapf(errortypes.ErrInvalidCoins, "cannot mint a non-positive amount")
+		return errorsmod.Wrapf(errortypes.ErrInvalidCoins, "cannot mint a non-positive amount")
 	}
 	_, err := sdk.AccAddressFromBech32(msg.Receiver)
 	if err != nil {
-		return sdkerrors.Wrap(err, "invalid receiver address")
+		return errorsmod.Wrap(err, "invalid receiver address")
 	}
 	if !common.IsHexAddress(msg.Sender) {
-		return sdkerrors.Wrapf(errortypes.ErrInvalidAddress, "invalid sender hex address %s", msg.Sender)
+		return errorsmod.Wrapf(errortypes.ErrInvalidAddress, "invalid sender hex address %s", msg.Sender)
 	}
 	return nil
 }